+++ core/core-api/src/main/kotlin/com/myongjiway/core/api/controller/AuthController.kt
@@ -1,6 +1,6 @@
-package com.myongjiway.core.api.auth.controller
+package com.myongjiway.core.api.controller
 
-import com.myongjiway.core.api.auth.controller.v1.request.KakaoLoginRequest
+import com.myongjiway.core.api.controller.v1.request.KakaoLoginRequest
 import com.myongjiway.core.api.controller.v1.request.RefreshRequest
 import com.myongjiway.core.api.support.response.ApiResponse
 import com.myongjiway.core.domain.auth.AuthService
