+++ core/core-domain/src/test/kotlin/com/myongjiway/core/domain/token/TokenTypeTest.kt
@@ -1,9 +1,5 @@
-package com.myongjiway.core.api.controller.v1.token
+package com.myongjiway.core.domain.token
 
-import com.myongjiway.core.domain.token.JwtProperty
-import com.myongjiway.core.domain.token.Token
-import com.myongjiway.core.domain.token.TokenGenerator
-import com.myongjiway.core.domain.token.TokenType
 import io.kotest.core.spec.style.FeatureSpec
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.types.shouldBeInstanceOf
